passing an empty hash probably means it's not a DSL
The current version of flog fails with a NoMethodError when attempting
to determine whether

    some_method({}) do |arg|
      # stuff
    end

is a dsl method or not. Flog's dsl_name? method currently makes the
assumption that when this code is parsed down into s-expressions that
the structure of a hash will always have a structure similar to the
following:

    s(:hash, s(:lit, :woot), s(:lit, 42)))

And therefore we can use the :[] methods on the sexp to check what the
contents of the hash are.

In our case the sexp is just

    s(:hash)

so when first_arg[1] is called we end up with `nil`, which we're not
expecting, we're expectin another sexp - s(:lit, :woot) in the above
example.

So when we try and get the elements of that sexp (which in our case is
nil) using [] - things go bang.

Our actual use case for this was flogging a file in a Rails project that
contained:

    respond_with({}, status: ok) do |format|
      format.html { redirect_to: somewhere_else }
    end

I have "fixed" this by returning false out of the dsl_name? method if
the hash passed as an argument is just an empty hash literal - I'm not
entirely sure whether returning true or false in this case is correct -
I made a call, so I could get the code running.